+++ src/test/java/lotto/model/user/UserLottoTest.java
@@ -2,7 +2,6 @@ package lotto.model.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.model.user.UserLotto;
 import lotto.util.generator.NumberGenerator;
 import lotto.util.generator.RandomNumberGenerator;
 import org.junit.jupiter.api.BeforeEach;
@@ -18,7 +17,7 @@ class UserLottoTest {
 
     @Test
     void 임의의_로또를_amount개_발행한다() {
-        int amount = 5;
+        long amount = 5L;
         UserLotto userLotto = new UserLotto(amount, numberGenerator);
         assertThat(userLotto.getLottos().size()).isEqualTo(amount);
     }
